+++ lib/materialist/materializer/internals/link_href_mapping.rb
@@ -9,8 +9,8 @@ module Materialist
         end
 
         def map(resource)
-          return unless link = resource.body._links[@key]
-          { @as => url_parser.call(link.href) }
+          return unless link = resource.dig(:_links, @key)
+          { @as => url_parser.call(link[:href]) }
         end
 
         private