+++ lib/materialist/materializer/internals/link_mapping.rb
@@ -2,25 +2,22 @@ module Materialist
   module Materializer
     module Internals
       class LinkMapping
-        def initialize(key:, enable_caching: false)
+        def initialize(key:, mapping: [], enable_caching: false)
           @key = key
+          @mapping = mapping
           @enable_caching = enable_caching
-          @mapping = []
         end
+
+        attr_reader :mapping
 
         def map(resource)
           return unless linked_resource = linked_resource(resource)
           mapping.map{ |m| m.map(linked_resource) }.compact.reduce(&:merge)
         end
 
-        attr_reader :mapping
-
-        private
-
         def linked_resource(resource)
-          return unless resource.body._links.include?(@key)
-          return unless sub_resource = resource.send(@key)
-          sub_resource.show(enable_caching: @enable_caching)
+          return unless href = resource.dig(:_links, @key, :href)
+          resource.client.get(href, options: { enable_caching: @enable_caching })
         rescue Routemaster::Errors::ResourceNotFound
           nil
         end