+++ lib/materialist/materializer_factory.rb
@@ -1,9 +1,7 @@
 module Materialist
   class MaterializerFactory
     def self.class_from_topic(topic)
-      "#{topic.to_s.singularize.classify}Materializer".constantize
-    rescue NameError
-      nil
+      "#{topic.to_s.singularize.classify}Materializer".safe_constantize
     end
   end
 end